restructure and pretest
