+++ lib/helper/request.js
@@ -1,5 +1,5 @@
 'use strict';
-const request = require('lib/actions/request');
+const request = require('request');
 const qs = require('querystring');
 const messages = require('elasticio-node').messages;
 const debug = require('debug')('webhook:request');
@@ -52,7 +52,9 @@ function buildRequestSettings(method, uri, secret) {
 function callback(err, response, body) {
     'use strict';
     if (err) {
+        // eslint-disable-next-line no-invalid-this
         this.emit('error', err);
+        // eslint-disable-next-line no-invalid-this
         this.emit('end');
         return;
     }
@@ -60,10 +62,13 @@ function callback(err, response, body) {
     const sc = response.statusCode;
 
     if (sc >= 200 && sc <= 206) {
+        // eslint-disable-next-line no-invalid-this
         this.emit('data', newMessage(response, body));
     } else {
+        // eslint-disable-next-line no-invalid-this
         this.emit('error', new Error('Endpoint responds with ' + sc));
     }
+    // eslint-disable-next-line no-invalid-this
     this.emit('end');
 }
