+++ lib/triggers/receive.js
@@ -1,6 +1,7 @@
 'use strict';
 const Q = require('q');
 
+// eslint-disable-next-line func-names
 exports.process = function (msg, conf) {
     const msgId = msg.id;
     console.log('Received new message with id', msgId);
@@ -29,7 +30,7 @@ exports.process = function (msg, conf) {
         console.log('Updated body', msg.body);
     }
 
-
+    // eslint-disable-next-line consistent-this
     let self = this;
 
     Q()
