+++ package.json
@@ -3,7 +3,7 @@
   "version": "1.1.0",
   "description": "elastic.io webhook component",
   "scripts": {
-    "pretest": "eslint spec lib/actions/get.js lib/actions/receive.js lib/actions/request.js lib/actions/send.js verifyCredentials.js --fix",
+    "pretest": "eslint spec lib verifyCredentials.js --fix",
     "test": "mocha spec --recursive"
   },
   "engines": {
@@ -13,18 +13,18 @@
   "license": "Apache License 2.0",
   "dependencies": {
     "debug": "3.1.0",
+    "elasticio-node": "0.0.9",
     "elasticio-sailor-nodejs": "2.4.2",
-    "elasticio-node": "0.0.8",
     "node-uuid": "1.4.8",
-    "request": "2.83.x",
-    "q": "1.5.1"
+    "q": "1.5.1",
+    "request": "2.83.x"
   },
   "devDependencies": {
+    "chai": "3.5.0",
     "eslint": "5.16.0",
     "eslint-plugin-mocha": "5.3.0",
     "mocha": "6.1.4",
     "nock": "10.0.6",
-    "sinon": "7.3.2",
-    "chai": "3.5.0"
+    "sinon": "7.3.2"
   }
 }
