+++ verifyCredentials.js
@@ -1,6 +1,5 @@
 'use strict';
 
-module.exports = verify;
 
 /**
  * Executes the verification logic by checking that fields are not empty using the provided apiKey.
@@ -14,7 +13,7 @@ const authTypes = {
     BASIC: 'BASIC'
 };
 
-function verify(credentials) {
+module.exports = function verify(credentials) {
     // access the value of the auth field defined in credentials section of component.json
     const { type, basic = {} } = credentials.auth;
 
@@ -31,4 +30,4 @@ function verify(credentials) {
     }
 
     return Promise.resolve(true);
-}
+};
